Add notes & fix for wasm
