+++ .gitignore
@@ -8,4 +8,8 @@ Cargo.lock
 .DS_Store
 
 /pasture-io/examples/out/test_output.las
-wasm-example
+
+render.d.ts
+render.js
+render_bg.wasm
+render_bg.wasm.d.ts
