--- pasture-tools/src/bin/render/wasm/render_bg.wasm.d.ts
@@ -1,65 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-export const memory: WebAssembly.Memory;
-export function main(a: number, b: number): number;
-export function wgpu_render_pass_set_pipeline(a: number, b: number): void;
-export function wgpu_render_pass_set_bind_group(a: number, b: number, c: number, d: number, e: number): void;
-export function wgpu_render_pass_set_vertex_buffer(a: number, b: number, c: number, d: number, e: number): void;
-export function wgpu_render_pass_set_push_constants(a: number, b: number, c: number, d: number, e: number): void;
-export function wgpu_render_pass_draw(a: number, b: number, c: number, d: number, e: number): void;
-export function wgpu_render_pass_draw_indexed(a: number, b: number, c: number, d: number, e: number, f: number): void;
-export function wgpu_render_pass_draw_indirect(a: number, b: number, c: number): void;
-export function wgpu_render_pass_draw_indexed_indirect(a: number, b: number, c: number): void;
-export function wgpu_render_pass_multi_draw_indirect(a: number, b: number, c: number, d: number): void;
-export function wgpu_render_pass_multi_draw_indexed_indirect(a: number, b: number, c: number, d: number): void;
-export function wgpu_render_pass_multi_draw_indirect_count(a: number, b: number, c: number, d: number, e: number, f: number): void;
-export function wgpu_render_pass_multi_draw_indexed_indirect_count(a: number, b: number, c: number, d: number, e: number, f: number): void;
-export function wgpu_render_pass_set_blend_constant(a: number, b: number): void;
-export function wgpu_render_pass_set_scissor_rect(a: number, b: number, c: number, d: number, e: number): void;
-export function wgpu_render_pass_set_viewport(a: number, b: number, c: number, d: number, e: number, f: number, g: number): void;
-export function wgpu_render_pass_set_stencil_reference(a: number, b: number): void;
-export function wgpu_render_pass_insert_debug_marker(a: number, b: number, c: number): void;
-export function wgpu_render_pass_push_debug_group(a: number, b: number, c: number): void;
-export function wgpu_render_pass_pop_debug_group(a: number): void;
-export function wgpu_render_pass_write_timestamp(a: number, b: number, c: number): void;
-export function wgpu_render_pass_begin_pipeline_statistics_query(a: number, b: number, c: number): void;
-export function wgpu_render_pass_end_pipeline_statistics_query(a: number): void;
-export function wgpu_render_bundle_set_pipeline(a: number, b: number): void;
-export function wgpu_render_bundle_set_bind_group(a: number, b: number, c: number, d: number, e: number): void;
-export function wgpu_render_bundle_set_vertex_buffer(a: number, b: number, c: number, d: number, e: number): void;
-export function wgpu_render_bundle_set_push_constants(a: number, b: number, c: number, d: number, e: number): void;
-export function wgpu_render_bundle_draw(a: number, b: number, c: number, d: number, e: number): void;
-export function wgpu_render_bundle_draw_indexed(a: number, b: number, c: number, d: number, e: number, f: number): void;
-export function wgpu_render_bundle_draw_indirect(a: number, b: number, c: number): void;
-export function wgpu_render_bundle_draw_indexed_indirect(a: number, b: number, c: number): void;
-export function wgpu_compute_pass_set_pipeline(a: number, b: number): void;
-export function wgpu_compute_pass_set_bind_group(a: number, b: number, c: number, d: number, e: number): void;
-export function wgpu_compute_pass_set_push_constant(a: number, b: number, c: number, d: number): void;
-export function wgpu_compute_pass_insert_debug_marker(a: number, b: number, c: number): void;
-export function wgpu_compute_pass_push_debug_group(a: number, b: number, c: number): void;
-export function wgpu_compute_pass_pop_debug_group(a: number): void;
-export function wgpu_compute_pass_write_timestamp(a: number, b: number, c: number): void;
-export function wgpu_compute_pass_begin_pipeline_statistics_query(a: number, b: number, c: number): void;
-export function wgpu_compute_pass_end_pipeline_statistics_query(a: number): void;
-export function wgpu_compute_pass_dispatch(a: number, b: number, c: number, d: number): void;
-export function wgpu_compute_pass_dispatch_indirect(a: number, b: number, c: number): void;
-export function wgpu_render_bundle_push_debug_group(a: number, b: number): void;
-export function wgpu_render_bundle_pop_debug_group(a: number): void;
-export function wgpu_render_bundle_insert_debug_marker(a: number, b: number): void;
-export function wgpu_render_pass_execute_bundles(a: number, b: number, c: number): void;
-export function __wbindgen_malloc(a: number): number;
-export function __wbindgen_realloc(a: number, b: number, c: number): number;
-export const __wbindgen_export_2: WebAssembly.Table;
-export function _dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h23c372c4f9c074d9(a: number, b: number, c: number): void;
-export function _dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__hdaa1bd42e9bba871(a: number, b: number, c: number): void;
-export function _dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h9b4f906d1e4563eb(a: number, b: number, c: number): void;
-export function _dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h0bafff2ffd06cabd(a: number, b: number, c: number): void;
-export function _dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h2adb89196b74f973(a: number, b: number, c: number): void;
-export function _dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h42f40be93671f73e(a: number, b: number, c: number): void;
-export function _dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h5f37a5660e0a1548(a: number, b: number, c: number): void;
-export function _dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__hbae10183ac0cebfe(a: number, b: number, c: number): void;
-export function _dyn_core__ops__function__FnMut_____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h973a8d0d3b4263b3(a: number, b: number): void;
-export function _dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__hd2b005e4759e8bc4(a: number, b: number, c: number): void;
-export function __wbindgen_free(a: number, b: number): void;
-export function __wbindgen_exn_store(a: number): void;
-export function __wbindgen_start(): void;
