--- pasture-tools/src/bin/tri.frag
@@ -1,8 +1,0 @@
-#version 450
-
-layout(location = 0) in vec3 inColor;
-layout(location = 0) out vec4 outFragColor;
-
-void main() {
-	outFragColor = vec4(inColor, 1.0);
-}
